% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{expand_with_id}
\alias{expand_with_id}
\title{Utility function to expand collapsed fields into tables}
\usage{
expand_with_id(x, id, id_col = "diarienummer", value_col = "values")
}
\arguments{
\item{x}{the string vector to flatten}

\item{id}{the string vector of corresponding identifiers}

\item{id_col}{string with title for identifier column, Default: 'diarienummer'}

\item{value_col}{string with the title for the values column, Default: 'values'}
}
\value{
table with flattened results
}
\description{
Utility function to expand collapsed fields into tables
}
