% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{formas_fields}
\alias{formas_fields}
\title{Description of fields included in responses}
\usage{
formas_fields()
}
\value{
a table with fields described
}
\description{
Description of fields included in responses
}
