% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{formas_lookup_tables}
\alias{formas_lookup_tables}
\title{Lookup tables for fields with composite values}
\usage{
formas_lookup_tables(x = formas_projects())
}
\arguments{
\item{x}{the formas projects data frame to "normalize", by default the output from formas_projects fcn.}
}
\value{
a list of data frames for fields with composite values
}
\description{
Some fields contain values separated by "; ". This function creates
separate lookup tables for these fields.
}
