% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{formas_project}
\alias{formas_project}
\title{Project details given one specific identifier}
\usage{
formas_project(id)
}
\arguments{
\item{id}{a identifier for the project, usually the value of the "diarienummer" field}
}
\value{
a table with results
}
\description{
Project details given one specific identifier
}
