% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{formas_projects_since}
\alias{formas_projects_since}
\title{Projects with changes since a specific date (inclusive)}
\usage{
formas_projects_since(from_date = Sys.Date() - 7)
}
\arguments{
\item{from_date}{a date, by default one week before the current date}
}
\value{
a table with results
}
\description{
Projects with changes since a specific date (inclusive)
}
