% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{rename_fields}
\alias{rename_fields}
\title{Rename fields to use English field names}
\usage{
rename_fields(x)
}
\arguments{
\item{x}{the data frame, usually resulting from a call to fcn formas_projects()}
}
\description{
These field names match closer to field names used in SweCRIS
}
